Added a script for processing raw rosbags and a script for analyzing the processed rosbags
the bash script add to a rosbag the static trasformation between the robodog tf  and the parent frame of the crazyflie embedded localizaion system ( magnetic included). Also it republish a new tf for the cf directly in the vicon/world frame.

The py script create plots and scores for each rosbags